pyproject.toml
[tool.poetry]
name = "agenthub-sdk"
version = "0.1.0"
description = "Development Toolkit for interacting with AgentHub's Open API"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.10.9"
pydantic = "^2.9.2"
python-socketio = "^5.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
